refactor: align explore tab hook with reviewer structure
